refactor: make CourseApi cross-platform using expect/actual HttpClient
