+++ shared/src/androidMain/kotlin/com/example/kmm/course/HttpClientProvider.kt
@@ -1,0 +1,17 @@
+package com.example.kmm.course
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.okhttp.OkHttp
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+actual class HttpClientProvider {
+    actual fun createClient(): HttpClient {
+        return HttpClient(OkHttp) {
+            install(ContentNegotiation) {
+                json(Json { ignoreUnknownKeys = true })
+            }
+        }
+    }
+}
