+++ shared/src/commonMain/kotlin/com/example/kmm/course/HttpClientProvider.kt
@@ -1,0 +1,7 @@
+package com.example.kmm.course
+
+import io.ktor.client.HttpClient
+
+expect class HttpClientProvider {
+    fun createClient(): HttpClient
+}
